+++ src/main/java/net/mysticrealms/fireworks/scavengerhunt/ScavengerListener.java
@@ -2,21 +2,24 @@ package net.mysticrealms.fireworks.scavengerhunt;
 
 import java.util.Map;
 
+import org.bukkit.ChatColor;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Player;
 import org.bukkit.entity.Projectile;
 import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.EntityDamageByEntityEvent;
 import org.bukkit.event.entity.EntityDamageEvent;
 import org.bukkit.event.entity.EntityDeathEvent;
+import org.bukkit.event.player.PlayerJoinEvent;
 
 public class ScavengerListener implements Listener {
 
 	private ScavengerHunt plugin;
 
-	public ScavengerListener(ScavengerHunt instance) {
-		plugin = instance;
+	public ScavengerListener(ScavengerHunt scavenger) {
+		plugin = scavenger;
 	}
 
 	@EventHandler
@@ -39,6 +42,19 @@ public class ScavengerListener implements Listener {
 			}
 			Map<EntityType, Integer> map = plugin.getMap(p.getName());
 			map.put(event.getEntity().getType(), map.get(event.getEntity().getType()) + 1);
+		}
+	}
+	
+	@EventHandler(priority = EventPriority.MONITOR)
+	public void onPlayerLogin(PlayerJoinEvent event) {
+		Player p = event.getPlayer();
+		long timeDiff = plugin.end - System.currentTimeMillis();
+		String timeLeft = ScavengerMessager.timeFormatter((int)((timeDiff)/1000));
+		if (plugin.isRunning) {
+			p.sendMessage(ChatColor.DARK_RED + "There is a Scavenger Hunt happening!");
+			p.sendMessage(ChatColor.DARK_RED + "You have: " + ChatColor.GOLD + timeLeft + "!");
+			p.sendMessage(ChatColor.DARK_RED + "Use " + ChatColor.GOLD + "/scavengerItems" + ChatColor.DARK_RED + " to view objectives.");
+			p.sendMessage(ChatColor.DARK_RED + "Use " + ChatColor.GOLD + "/scavengerRewards" + ChatColor.DARK_RED + " to view rewards.");
 		}
 	}
 }