+++ src/main/java/net/mysticrealms/fireworks/scavengerhunt/util/NameRetriever.java
@@ -15,8 +15,9 @@ import java.util.Map;
  import java.util.TreeMap;
  */
 
-public class PotionRetriever {
+public class NameRetriever {
 	public static Map<Short, String> potionMap = new HashMap<Short, String>();
+	public static Map<Integer, String> discMap = new HashMap<Integer, String>();
 
 	public static void setPotionMap() {
 		String formatter = "Potion of ";
@@ -34,8 +35,7 @@ public class PotionRetriever {
 		potionMap.put((short) 8297, formatter + "Strength II (Unbrewable)");
 		potionMap.put((short) 8292, formatter + "Poison II (Unbrewable)");
 
-		// effective potions. There are some skipped DV's, will try to cleanup
-		// later
+		// effective potions. There are some skipped DV's, will try to cleanup later
 		String potionType[] = new String[] { "Regeneration", "Swiftness", "Fire Resistance", "Poison", "Instant Health", "Weakness", "Strength", "Slowness", "Harming" };
 		for (short i = 8193; i < 8193 + potionType.length - 4; i++) {
 			potionMap.put(i, formatter + potionType[i - 8193]);
@@ -78,15 +78,37 @@ public class PotionRetriever {
 		potionMap.put((short) 16460, "Splash " + formatter + "Harming (Reverted)");
 	}
 
-	public static String getPotionName(short potionDV) {
-		String potionName = null;
-		potionName = potionMap.get(potionDV).toString();
-		return potionName;
+	public static void setDiscMap() {
+		String formatter = "Music Disc - ";
+
+		discMap.put(2256, formatter + "13");
+		discMap.put(2257, formatter + "cat");
+		discMap.put(2258, formatter + "blocks");
+		discMap.put(2259, formatter + "chirp");
+		discMap.put(2260, formatter + "far");
+		discMap.put(2261, formatter + "mall");
+		discMap.put(2262, formatter + "mellohi");
+		discMap.put(2263, formatter + "stal");
+		discMap.put(2264, formatter + "strad");
+		discMap.put(2265, formatter + "ward");
+		discMap.put(2266, formatter + "11");
 	}
 
-	/*
-	 * Prints potion map (debug) public static void printPotionMap() { List sortedKeys=new ArrayList(potionMap.keySet());
-	 * Collections.sort(sortedKeys); Map<Short, String> treeMap = new TreeMap<Short, String>(potionMap); for (short key : treeMap.keySet()) {
-	 * System.out.println(key+":"+potionMap.get(key)); } }
-	 */
+	public static String getPotionName(short potionDV) {
+		return potionMap.get(potionDV).toString();
+	}
+
+	public static String getDiscName(int discDV) {
+		return discMap.get(discDV).toString();
+	}
+
+	// Prints potion map (debug)
+	// public static void printPotionMap() {
+	// List sortedKeys=new ArrayList(potionMap.keySet());
+	// Collections.sort(sortedKeys);
+	// Map<Short, String> treeMap = new TreeMap<Short, String>(potionMap);
+	// for (short key : treeMap.keySet()) {
+	// System.out.println(key+":"+potionMap.get(key));
+	// }
+	// }
 }